refactor: Group chatroom components
