+++ src/components/ChatRoom/ChatBubble.tsx
